QuizCards
LoadQuizCards

+++ RtgVsZmbs/Data/QuizAnswer.cs
@@ -8,7 +8,7 @@ namespace RtgVsZmbs.Objects
 {
     public class QuizAnswer
     {
-        public Int64 ID { get; private set; }
+        public Int32 ID { get; private set; }
 
         public int TypeID { get; private set; }
 
@@ -16,7 +16,7 @@ namespace RtgVsZmbs.Objects
 
         public Boolean IsCorrect { get; set; }
 
-        public QuizAnswer(Int64 id, string answer, Boolean isCorrect, int typeid)
+        public QuizAnswer(int id, string answer, Boolean isCorrect, int typeid)
         {
             ID = id;
             Answer = answer;
